requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annlite==0.5.9
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
dataclasses-json==0.5.7
dnspython==2.3.0
docarray==0.21.0
docker==6.0.1
ecdsa==0.18.0
fastapi==0.95.1
filelock==3.11.0
filesplit==4.0.1
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
gptcache==0.1.11
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
jina-hubble-sdk==0.36.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.141
loguru==0.7.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pinecone-client==2.2.1
protobuf==4.22.3
pyasn1==0.4.8
pydantic==1.10.7
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
rocksdict==0.3.11
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
